+++ backend/package.json
@@ -17,6 +17,7 @@
     "dotenv": "^16.3.1",
     "express": "^4.17.3",
     "express-async-handler": "^1.2.0",
+    "express-list-endpoints": "^6.0.0",
     "mongodb": "^6.3.0",
     "mongoose": "^8.0.0",
     "nodemon": "^3.0.1"
